+++ PatheticMain/src/main/java/xyz/oli/pathing/bstats/BStatsHandler.java
@@ -1,4 +1,4 @@
-package xyz.oli.pathing.model.finder.bstats;
+package xyz.oli.pathing.bstats;
 
 import lombok.experimental.UtilityClass;
 import org.bukkit.plugin.java.JavaPlugin;
